Fixing concerns/term files
